CLIENT-GAME: Add processor directive
MINOR
