+++ src/Client/Systems/Graphic/ParallaxSystems.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <cstddef>
+#include <functional>
+#include <vector>
+
+namespace Systems::ParallaxSystems {
+    std::vector<std::function<void(std::size_t, std::size_t)>> getParallaxSystems();
+}
